requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pinecone-client>=5.0.0
pinecone-plugin-assistant
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-pinecone
python-multipart>=0.0.6
pydantic>=2.5.0
google-generativeai

